Rename experiment.ts to test.ts
The CLI command is test. Keep the module name the same.
